ForeignKey Query after inserting products (excel/csv file) to sql server
This sql query should be added to establish foreign key because when the products csv file got imported to sql server, the foreign key was not there. 
This is important because when the application runs,  the results in Index page should include the product category which requires the foreign key relationship to detect the category number for each product.